[Database] add triggers to G/L Account table to handle parent Accounts
Added before insert and before update triggers to automatically set
a reference to the parent table based on the no field.

Tried to enforce as much of the data related business rules at
database level as possible.

Closes #3.[Database] add triggers to G/L Account table to handle parent
Accounts

Added before insert and before update triggers to automatically set
a reference to the parent table based on the no field.

Tried to enforce as much of the data related business rules at
database level as possible.

Closes #3.[Database] add triggers to G/L Account table to handle parent
Accounts

Added before insert and before update triggers to automatically set
a reference to the parent table based on the no field.

Tried to enforce as much of the data related business rules at
database level as possible.

Closes #3.[Database] add triggers to G/L Account table to handle parent
Accounts

Added before insert and before update triggers to automatically set
a reference to the parent table based on the no field.

Tried to enforce as much of the data related business rules at
database level as possible.

Closes #3.[Database] add triggers to G/L Account table to handle parent
Accounts

Added before insert and before update triggers to automatically set
a reference to the parent table based on the no field.

Tried to enforce as much of the data related business rules at
database level as possible.

Closes #3.[Database] add triggers to G/L Account table to handle parent
Accounts

Added before insert and before update triggers to automatically set
a reference to the parent table based on the no field.

Tried to enforce as much of the data related business rules at
database level as possible.

Closes #3.[Database] add triggers to G/L Account table to handle parent
Accounts

Added before insert and before update triggers to automatically set
a reference to the parent table based on the no field.

Tried to enforce as much of the data related business rules at
database level as possible.

Closes #3.[Database] add triggers to G/L Account table to handle parent
Accounts

Added before insert and before update triggers to automatically set
a reference to the parent table based on the no field.

Tried to enforce as much of the data related business rules at
database level as possible.

Closes #3.